Add csv output generation and move code to "aggregate_api"
