+++ .gitignore
@@ -1,2 +1,3 @@
 
 .DS_Store
+test_aggregate_data.csv
