+++ sdk/aqualink_sdk/api/time_series_api.py
@@ -23,8 +23,6 @@ from aqualink_sdk.model_utils import (  # noqa: F401
 )
 from aqualink_sdk.model.inline_response200 import InlineResponse200
 from aqualink_sdk.model.inline_response2001 import InlineResponse2001
-from aqualink_sdk.api.aggregate import aggregate_data
-
 
 class TimeSeriesApi(object):
     """NOTE: This class is auto generated by OpenAPI Generator
@@ -447,82 +445,6 @@ class TimeSeriesApi(object):
         kwargs['hourly'] = \
             hourly
         return self.time_series_controller_find_site_data_endpoint.call_with_http_info(**kwargs)
-
-    def time_series_controller_find_aggregate_site_data(
-        self,
-        site_id,
-        metrics,
-        start,
-        end,
-        aggregate_frequency,
-        aggregate_mapping,
-        **kwargs
-    ):
-        """Returns specified time series data for a specified site  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.time_series_controller_find_site_data(site_id, metrics, start, end, hourly, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            site_id (float):
-            metrics ([str]):
-            start (str):
-            end (str):
-            hourly (bool):
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _spec_property_naming (bool): True if the variable names in the input data
-                are serialized names, as specified in the OpenAPI document.
-                False if the variable names in the input data
-                are pythonic names, e.g. snake case (default)
-            _content_type (str/None): force body content-type.
-                Default is None and content-type will be predicted by allowed
-                content-types and body.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            InlineResponse200
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['hourly'] = False
-        api_response = self.time_series_controller_find_site_data(
-            site_id,
-            metrics,
-            start,
-            end,
-            async_req=False,
-            **kwargs
-        )
-
-        if aggregate_frequency and aggregate_mapping:
-            # TODO - verify that all the metrics requested in "metrics" have a mapping
-            # TODO - strongly type the frquency
-            # TODO - Hourly and aggregate are incompatible
-            data = aggregate_data(api_response, aggregate_frequency, aggregate_mapping)
-            api_response._data_store = data
-        return api_response
 
 
     def time_series_controller_find_site_data_range(
