+++ sdk/setup.py
@@ -3,7 +3,7 @@
 
     The Aqualink public API documentation  # noqa: E501
 
-    The version of the OpenAPI document: 1.0.0
+    The version of the OpenAPI document: 1.0.4
     Generated by: https://openapi-generator.tech
 """
 
@@ -11,7 +11,7 @@
 from setuptools import setup, find_packages  # noqa: H301
 
 NAME = "aqualink-api"
-VERSION = "1.0.0"
+VERSION = "1.0.4"
 # To install the library, run the following
 #
 # python setup.py install
